Credit Card Defaults

A dataset containing information on default payments, demographic infomration, and payment history for Taiwanese credit card clients from April 2005 to Septembrer 2005

CCD

Format

A dataset with 29728 rows and 25 variables. The variables are as follows:

default.payment.next.month

Factor indicating default

LIMIT

Credit card limit

SEX

Factor indicating gender

Education

Factor depicting highest level of education

Marriage

Factor indicating marital status

AGE

Age

REPAID.SEPT

Factor describing September payment status

REPAID.AUG

Factor describing August payment status

REPAID.JULY

Factor describing July payment status

REPAID.JUNE

Factor describing June payment status

REPAID.MAY

Factor describing May payment status

REPAID.APR

Factor describing April payment status

BILL_SEPT

Total credit card bill in September

BILL_AUG

Total credit card bill in August

BILL_JULY

Total credit card bill in July

BILL_JUNE

Total credit card bill in June

BILL_MAY

Total credit card bill in May

BILL_APR

Total credit card bill in April

PAYMENT_SEPT

Total paid in September

PAYMENT_AUG

Total paid in August

PAYMENT_JULY

Total paid in July

PAYMENT_JUNE

Total paid in June

PAYMENT_MAY

Total paid in May

PAYMENT_APR

Total paid in April

@source Lichman, M. (2013). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.

Details

This dataset was retrieved by UCI's Machine Learning Repository and then slightly modified to improve the predictiveness of the variables.